Separate User, UserSite table from company.py
