+++ app/db/base.py
@@ -8,16 +8,18 @@ from app.models.enums import CycleStatusEnum
 from app.models.enums import AttachmentTypeEnum
 from app.models.enums import ReviewStatusEnum
 
-# -- attachment 관련 모델
-from app.models.attachment import EstimateAttachment
-from app.models.attachment import EstimateAttachmentInfo
-from app.models.attachment import EstimateAttachmentReview
-
 # -- company 관련 모델
 from app.models.company import Company
 from app.models.company import Site
-from app.models.company import User
-from app.models.company import UserSite
+
+# -- user 관련 모델
+from app.models.user import User
+from app.models.user import UserSite
+
+# -- project_code 관련 모델
+from app.models.project_code import ProjectCode
+from app.models.project_code import VendorCode
+from app.models.project_code import ProjectCodeClient
 
 # -- estimate 관련 모델
 from app.models.estimate import EstimateInfo
@@ -26,10 +28,11 @@ from app.models.estimate import EstimateFeedback
 from app.models.estimate import EstimateResult
 from app.models.estimate import EstimateInfoShare
 
-# -- project_code 관련 모델
-from app.models.project_code import ProjectCode
-from app.models.project_code import VendorCode
-from app.models.project_code import ProjectCodeClient
+# -- attachment 관련 모델
+from app.models.attachment import EstimateAttachment
+from app.models.attachment import EstimateAttachmentInfo
+from app.models.attachment import EstimateAttachmentReview
+
 
 __all__ = [
     "Base",
