+++ app/models/__init__.py
@@ -1,5 +1,5 @@
 from .enums import *
-from .attachment import *
 from .company import *
+from .user import *
 from .estimate import *
 from .project_code import *
